# mnt/private/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de l'API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API de développement"
        }
    ],
    "paths": {
...
        "/api/session-types": {
            "get": {
                "tags": [
                    "Session Types"
                ],
                "summary": "Retrieve all session types",
                "operationId": "d3259be72b185a4fc96dcc882f335a73",
                "responses": {
                    "200": {
                        "description": "List of all session types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SessionType"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Session Types"
                ],
                "summary": "Create a new session type",
                "operationId": "f760f8a1896d7201c06e6fcd551ea573",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "SESSTYPE_LABEL"
                                ],
                                "properties": {
                                    "SESSTYPE_LABEL": {
                                        "type": "string",
                                        "example": "Theoretical training"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session Type Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionType"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/session-types/{id}": {
            "get": {
                "tags": [
                    "Session Types"
                ],
                "summary": "Retrieve a specific session type by ID",
                "operationId": "63a95297f0f46da5f744a8a7cbeecf79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session type found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session type not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Session Types"
                ],
                "summary": "Update an existing session type",
                "operationId": "286a242819b71a01826f0dbae000dff5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "SESSTYPE_LABEL"
                                ],
                                "properties": {
                                    "SESSTYPE_LABEL": {
                                        "type": "string",
                                        "example": "Practical training"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session type updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session type not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Session Types"
                ],
                "summary": "Delete a session type",
                "operationId": "98ae20cadf3a2211a1277f557c5b6306",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session type deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Session type deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session type not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SessionType": {
                "required": [
                    "SESSTYPE_LABEL"
                ],
                "properties": {
                    "id": {
                        "description": "Session Type ID",
                        "type": "integer"
                    },
                    "SESSTYPE_LABEL": {
                        "description": "Session Type Label",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}